+++ angular-workspace/components/highlight-box/highlight-box.stories.ts
@@ -1,5 +1,5 @@
 import { type Meta, moduleMetadata } from "@storybook/angular";
-import { HighlightBoxArgs, highlightBoxMeta, highlightBoxStories } from "dso-toolkit";
+import { HighlightBoxArgs, HighlightBoxColor, highlightBoxMeta, highlightBoxStories } from "dso-toolkit";
 
 import { DsoHighlightBox } from "../../projects/component-library/src/public-api";
 import { templateContainer } from "../../templates";
@@ -15,11 +15,14 @@ const meta: Meta<HighlightBoxArgs> = {
     }),
   ],
   title: "Highlight Box",
+  args: {
+    color: HighlightBoxColor.grey,
+  },
 };
 
 export default meta;
 
-const { Default, Yellow, WhiteWithDropshadow, WithBorder, WithIcon, WithBannerImage } = highlightBoxStories({
+const { Default, Yellow, Green, WhiteWithDropshadow, WithBorder, WithIcon, WithBannerImage } = highlightBoxStories({
   templateContainer,
   storyTemplates: (templates) => {
     const { highlightBoxTemplate } = templates;
@@ -31,4 +34,4 @@ const { Default, Yellow, WhiteWithDropshadow, WithBorder, WithIcon, WithBannerIm
   },
 });
 
-export { Default, WhiteWithDropshadow, WithBannerImage, WithBorder, WithIcon, Yellow };
+export { Default, Green, WhiteWithDropshadow, WithBannerImage, WithBorder, WithIcon, Yellow };