+++ packages/core/src/components/highlight-box/highlight-box.tsx
@@ -1,4 +1,5 @@
 import { Component, Element, Host, Prop, h } from "@stencil/core";
+import { HighlightBoxColor } from "dso-toolkit";
 
 @Component({
   tag: "dso-highlight-box",
@@ -7,22 +8,16 @@ import { Component, Element, Host, Prop, h } from "@stencil/core";
 })
 export class HighlightBox {
   /**
-   * For yellow Highlight Box.
+   * The color of the Highlight Box. Default is "grey".
    */
   @Prop({ reflect: true })
-  yellow?: boolean;
+  color: string = HighlightBoxColor.grey;
 
   /**
    * For a bordered Highlight Box.
    */
   @Prop({ reflect: true })
   border?: boolean;
-
-  /**
-   * For a while Highlight Box.
-   */
-  @Prop({ reflect: true })
-  white?: boolean;
 
   /**
    * For an Highlight Box with a drop shadow.