+++ packages/dso-toolkit/src/components/highlight-box/highlight-box.models.ts
@@ -1,7 +1,14 @@
+export const HighlightBoxColor = {
+  yellow: "yellow",
+  white: "white",
+  grey: "grey",
+  green: "green",
+} as const;
+
+export type HighlightBoxColor = (typeof HighlightBoxColor)[keyof typeof HighlightBoxColor];
+
 export interface HighlightBox<TemplateFnReturnType = string> {
-  yellow?: boolean;
-  white?: boolean;
-  grey?: boolean;
+  color: HighlightBoxColor;
   dropShadow?: boolean;
   border?: boolean;
   step?: number;