+++ storybook/src/components/highlight-box/highlight-box.core-template.ts
@@ -1,4 +1,4 @@
-import { HighlightBox } from "dso-toolkit";
+import { HighlightBox, HighlightBoxColor } from "dso-toolkit";
 import { html, nothing } from "lit-html";
 import { ifDefined } from "lit-html/directives/if-defined.js";
 import { unsafeHTML } from "lit-html/directives/unsafe-html.js";
@@ -9,13 +9,11 @@ export const coreHighlightBox: ComponentImplementation<HighlightBox> = {
   component: "highlightBox",
   implementation: "core",
   template: ({ iconTemplate, imageTemplate }) =>
-    function highlightBoxTemplate({ yellow, white, grey, dropShadow, border, step, icon, content, bannerImage }) {
+    function highlightBoxTemplate({ color, dropShadow, border, step, icon, content, bannerImage }) {
       return html`
         <dso-highlight-box
           step=${ifDefined(typeof step === "number" && step > 0 ? step : undefined)}
-          ?yellow=${yellow}
-          ?white=${white}
-          ?grey=${grey}
+          color=${ifDefined(color ?? HighlightBoxColor.grey)}
           ?drop-shadow=${dropShadow}
           ?border=${border}
         >